linter fixes.

+++ client/vue/penandpaper/src/rendering/material.ts
@@ -1,17 +1,16 @@
 import Camera from "./camera"
 import ShaderCache from "./shader_cache"
-import Actor from "./actor"
+import Actor, { ShaderInputType } from './actor'
 
 export default class Material {
   vertexShaderSrc: string = ''
   fragmentShaderSrc: string = ''
 
-  program?: WebGLProgram 
+  program: WebGLProgram | null = null
 
   isBuilt: boolean = false
 
-  // TODO: define a system of vertex attributes
-  positionAttr: number = -1
+  attributes: Map<ShaderInputType, number> = new Map()
 
   build (ctx: WebGLRenderingContext, shaderCache: ShaderCache) {
     // TODO: cache the program and reuse it where possible
@@ -23,15 +22,19 @@ export default class Material {
       let vertexShader = this.buildShader(ctx, this.vertexShaderSrc, ctx.VERTEX_SHADER)
       let fragmentShader = this.buildShader(ctx, this.fragmentShaderSrc, ctx.FRAGMENT_SHADER)
 
-      this.program = ctx.createProgram()
-      ctx.attachShader(this.program, vertexShader)
-      ctx.attachShader(this.program, fragmentShader)
-      ctx.linkProgram(this.program)
+      if (vertexShader && fragmentShader) {
+        this.program = ctx.createProgram()
+        if (this.program) {
+          ctx.attachShader(this.program, vertexShader)
+          ctx.attachShader(this.program, fragmentShader)
+          ctx.linkProgram(this.program)
 
-      if (!ctx.getProgramParameter(this.program, ctx.LINK_STATUS)) {
-        console.log('Error: unable to link the shader: ', ctx.getProgramInfoLog(this.program))
+          if (!ctx.getProgramParameter(this.program, ctx.LINK_STATUS)) {
+            console.log('Error: unable to link the shader: ', ctx.getProgramInfoLog(this.program))
+          }
+          shaderCache.addProgram(this.vertexShaderSrc, this.fragmentShaderSrc, this.program)
+        }
       }
-      shaderCache.addProgram(this.vertexShaderSrc, this.fragmentShaderSrc, this.program)
     }
   }
 
@@ -39,16 +42,20 @@ export default class Material {
     if (!this.isBuilt) {
       this.build(ctx, shaderCache)
     }
-    ctx.useProgram(this.program)
+    if (this.program) {
+      ctx.useProgram(this.program)
+    }
   }
 
-  buildShader (ctx: WebGLRenderingContext, src: string, type: number) : WebGLShader{
+  buildShader (ctx: WebGLRenderingContext, src: string, type: number) : WebGLShader | null {
     let shader = ctx.createShader(type)
-    ctx.shaderSource(shader, src)
-    ctx.compileShader(shader)
-    if (!ctx.getShaderParameter(shader, ctx.COMPILE_STATUS)) {
-      console.log('Error: unable to compile the shader: ', src, ctx.getShaderInfoLog(shader))
-      ctx.deleteShader(shader)
+    if (shader) {
+      ctx.shaderSource(shader, src)
+      ctx.compileShader(shader)
+      if (!ctx.getShaderParameter(shader, ctx.COMPILE_STATUS)) {
+        console.log('Error: unable to compile the shader: ', src, ctx.getShaderInfoLog(shader))
+        ctx.deleteShader(shader)
+      }
     }
     return shader
   }
